Create custom hook for using pads
