+++ src/App.js
@@ -3,17 +3,11 @@ import { BrowserRouter as Router } from 'react-router-dom'
 import styled from 'styled-components/macro'
 import { padsData } from './common/padsData'
 import SamplePlayer from './common/SamplePlayer'
-import {
-  getKeyByName,
-  getPadNameByKey,
-  focusElementAfterLoad,
-  stopPinchZooming,
-  toLowerNoWhiteSpace,
-  updateInArray,
-} from './common/utils'
+import { focusElementAfterLoad } from './common/utils'
 import Display from './components/Display'
 import DisplayNav from './components/DisplayNav'
 import PadSection from './components/Pad/PadSection'
+import usePads from './hooks/usePads'
 
 const samplePlayer = new SamplePlayer(padsData)
 
@@ -21,8 +15,14 @@ export default function App() {
   const [masterVolume, setMasterVolume] = useState(
     samplePlayer.getMasterVolume()
   )
-  const [pads, setPads] = useState(padsData)
-  const [selectedPad, setSelectedPad] = useState(null)
+  const {
+    pads,
+    selectedPad,
+    handleKeyDown,
+    handleKeyUp,
+    handlePadTouchStart,
+    handlePadTouchEnd,
+  } = usePads(samplePlayer)
   const appElement = useRef(null)
   focusElementAfterLoad(appElement)
 
@@ -30,8 +30,8 @@ export default function App() {
     <Router>
       <AppStyled
         ref={appElement}
-        onKeyDown={onKeyDown}
-        onKeyUp={onKeyUp}
+        onKeyDown={handleKeyDown}
+        onKeyUp={handleKeyUp}
         onContextMenu={event => {
           event.preventDefault()
         }}
@@ -60,47 +60,6 @@ export default function App() {
     const volumeInDB = targetValue > 6 ? 6 : targetValue
     setMasterVolume(volumeInDB)
     samplePlayer.setMasterVolume(volumeInDB)
-  }
-
-  function onKeyDown(event) {
-    if (event.repeat) {
-      return
-    }
-    const key = event.key
-    samplePlayer.playSample(key)
-    setPadIsTriggered(key, true)
-    const padName = getPadNameByKey(pads, key)
-    setSelectedPad(padName)
-  }
-
-  function onKeyUp(event) {
-    const key = event.key
-    setPadIsTriggered(key, false)
-  }
-
-  function handlePadTouchStart(padName) {
-    const name = toLowerNoWhiteSpace(padName)
-    const key = getKeyByName(pads, name)
-    samplePlayer.playSample(key)
-    setPadIsTriggered(key, true)
-    setSelectedPad(padName)
-  }
-
-  function handlePadTouchEnd(event, padName) {
-    stopPinchZooming(event)
-    const name = toLowerNoWhiteSpace(padName)
-    const key = getKeyByName(pads, name)
-    setPadIsTriggered(key, false)
-  }
-
-  function setPadIsTriggered(key, triggerd) {
-    const index = pads.findIndex(pad => pad.key === key)
-    if (index > -1) {
-      const pad = { ...pads[index] }
-      pad.isTriggered = triggerd
-      const newPads = updateInArray(pads, pad, index)
-      setPads(newPads)
-    }
   }
 }
